extract precompile-error crate
