+++ sdk/precompile-error/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "solana-precompile-error"
+description = "Solana PrecompileError type"
+documentation = "https://docs.rs/solana-precompile-error"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+solana-decode-error = { workspace = true }
+num-derive = { workspace = true }
+num-traits = { workspace = true }
+thiserror = { workspace = true }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
